SNMP-BRIDGE/requirements.txt
elasticsearch>=8.0.0
pyyaml>=6.0
requests>=2.28.0
prometheus-client>=0.16.0
python-dotenv>=1.0.0
argparse>=1.4.0
pytest>=7.0.0
pytest-cov>=4.0.0
